+++ packages/deployment-service/src/entities/pipeline.entity.ts
@@ -1,30 +1,10 @@
-import { AppTypeEnum, PackageManagerEnum, PipelineModelInterface } from '@reapit/foundations-ts-definitions'
+import { AppTypeEnum, PackageManagerEnum, pipelineDeploymentDisabled, PipelineModelInterface, pipelineNotDeletable, PipelineBuildStatus } from '@reapit/foundations-ts-definitions'
 import { Column, Entity, ManyToOne, OneToMany } from 'typeorm'
 import { AbstractEntity } from './abstract-entity'
 import { PipelineRunnerEntity } from './pipeline-runner.entity'
-import { PipelineBuildStatus } from '../pipeline/pipeline-dto'
 import { BitbucketClientEntity } from './bitbucket-client.entity'
 import { Exclude, Type } from 'class-transformer'
 
-export const pipelineDeploymentDisabled = [
-  'PROVISIONING',
-  'PROVISION_REQUEST',
-  'FAILED_TO_PROVISION',
-  'PRE_PROVISIONED',
-  'DELETING',
-  'DELETED',
-  'SCHEDULED_FOR_DELETION',
-  'DELETION_REQUEST',
-]
-export const pipelineNotDeletable = [
-  'IN_PROGRESS',
-  'DELETING',
-  'PROVISION_REQUEST',
-  'PROVISIONING',
-  'QUEUED',
-  'SCHEDULED_FOR_DELETION',
-  'DELETION_REQUEST',
-]
 @Entity('pipelines')
 export class PipelineEntity extends AbstractEntity implements PipelineModelInterface {
   @Column()
