+++ packages/deployment-service/src/pipeline/pipeline-dto.ts
@@ -1,5 +1,4 @@
 import { IsString, IsNotEmpty, IsOptional, IsEnum, MaxLength, IsUUID, Matches, IsUrl } from 'class-validator'
-import { CodeBuild } from 'aws-sdk'
 
 export enum AppTypeEnum {
   REACT = 'react',
@@ -10,24 +9,6 @@ export enum PackageManagerEnum {
   YARN = 'yarn',
   NPM = 'npm',
 }
-
-export type PipelineProvisionBuildStatuses =
-  | 'PRE_PROVISIONED'
-  | 'PROVISIONING'
-  | 'PROVISION_REQUEST'
-  | 'FAILED_TO_PROVISION'
-
-export type PipelineDeployingBuildStatues = 'QUEUED' | 'COMPLETED' | 'FAILED'
-
-export type PipelineDeleteBuildStatuses = 'DELETED' | 'DELETING' | 'SCHEDULED_FOR_DELETION' | 'DELETION_REQUEST'
-
-export type PipelineBuildStatus =
-  | 'CREATED'
-  | ('READY_FOR_DEPLOYMENT' &
-      PipelineDeleteBuildStatuses &
-      PipelineDeployingBuildStatues &
-      PipelineProvisionBuildStatuses)
-  | CodeBuild.StatusType
 
 export class PipelineDto {
   @IsString()
